Make AI work with other programs
This change makes AI call decode(x,y,[d/r]) before going into update(disp,dir) and allows for the program to repeat if it gets a point in update(). Along with this, t4 does not have to stay constant across calls (because I remembered how to properly put $ra on the stack).